--- Sources/HTTPRouting/Routing+HTTP.swift
@@ -1,8 +1,0 @@
-import HTTP
-import Routing
-
-/**
-    The default Router for HTTP
-    which routes paths to Responders.
-*/
-//public typealias Router = HTTPRouter
