+++ Sources/HTTPRouting/Utilities.swift
@@ -1,9 +1,7 @@
 extension String {
-    /**
-        Separates a URI path into
-        an array by splitting on `/`
-    */
-    var pathComponents: [String] {
+    /// Separates a URI path into
+    /// an array by splitting on `/`
+    internal var pathComponents: [String] {
         return characters
             .split(separator: "/", omittingEmptySubsequences: true)
             .map { String($0) }
@@ -11,7 +9,10 @@ extension String {
 }
 
 extension Sequence where Iterator.Element == String {
-    var pathComponents: [String] {
-        return flatMap { $0.pathComponents }
+    /// Ensures that `/` are interpreted properly on arrays
+    /// of path components, so `["foo", "bar/dar"]`
+    /// will expand to `["foo", "bar", "dar"]`
+    internal var pathComponents: [String] {
+        return flatMap { $0.pathComponents } .filter { !$0.isEmpty }
     }
 }
